fix base os
